#191 remove superfluous layer SystemMetrics
