+++ src/main/kotlin/com/krillsson/sysapi/core/metrics/Metrics.kt
@@ -1,5 +1,9 @@
 package com.krillsson.sysapi.core.metrics
 
+import com.krillsson.sysapi.core.domain.processes.ProcessSort
+import com.krillsson.sysapi.core.domain.system.SystemInfo
+import com.krillsson.sysapi.core.domain.system.SystemLoad
+
 interface Metrics {
     fun initialize()
     fun cpuMetrics(): CpuMetrics
@@ -10,5 +14,10 @@ interface Metrics {
     fun processesMetrics(): ProcessesMetrics
     fun gpuMetrics(): GpuMetrics
     fun motherboardMetrics(): MotherboardMetrics
-    fun systemMetrics(): SystemMetrics
+    fun systemLoad(
+        sort: ProcessSort = ProcessSort.MEMORY,
+        limit: Int = -1
+    ): SystemLoad
+
+    fun systemInfo(): SystemInfo
 }
