--- src/main/kotlin/com/krillsson/sysapi/core/metrics/cache/CachingSystemMetrics.kt
@@ -1,29 +1,0 @@
-package com.krillsson.sysapi.core.metrics.cache
-
-import com.krillsson.sysapi.core.domain.system.OperatingSystem
-import com.krillsson.sysapi.core.domain.system.Platform
-import com.krillsson.sysapi.core.metrics.defaultimpl.DefaultSystemMetrics
-
-class CachingSystemMetrics(
-    cpuMetrics: CachingCpuMetrics,
-    networkMetrics: CachingNetworkMetrics,
-    diskMetrics: CachingDiskMetrics,
-    fileSystemMetrics: CachingFileSystemMetrics,
-    memoryMetrics: CachingMemoryMetrics,
-    processesMetrics: CachingProcessesMetrics,
-    motherboardMetrics: CachingMotherboardMetrics,
-    gpuMetrics: CachingGpuMetrics,
-    platform: Platform,
-    operatingSystem: OperatingSystem
-) : DefaultSystemMetrics(
-    cpuMetrics,
-    networkMetrics,
-    diskMetrics,
-    fileSystemMetrics,
-    memoryMetrics,
-    processesMetrics,
-    motherboardMetrics,
-    gpuMetrics,
-    operatingSystem,
-    platform
-)
