+++ src/main/kotlin/com/krillsson/sysapi/core/metrics/defaultimpl/DefaultMetrics.kt
@@ -1,6 +1,12 @@
 package com.krillsson.sysapi.core.metrics.defaultimpl
 
+import com.krillsson.sysapi.core.domain.processes.ProcessSort
+import com.krillsson.sysapi.core.domain.system.OperatingSystem
+import com.krillsson.sysapi.core.domain.system.Platform
+import com.krillsson.sysapi.core.domain.system.SystemInfo
+import com.krillsson.sysapi.core.domain.system.SystemLoad
 import com.krillsson.sysapi.core.metrics.*
+import com.krillsson.sysapi.util.EnvironmentUtils
 import org.springframework.beans.factory.annotation.Qualifier
 import org.springframework.stereotype.Component
 
@@ -14,7 +20,8 @@ open class DefaultMetrics(
     private val processesMetrics: DefaultProcessesMetrics,
     @Qualifier("defaultMotherboardMetrics") private val motherboardMetrics: MotherboardMetrics,
     private val memoryMetrics: MemoryMetrics,
-    private val systemMetrics: SystemMetrics,
+    private val operatingSystem: OperatingSystem,
+    private val platform: Platform
 ) : Metrics {
 
     override fun initialize() {
@@ -54,7 +61,33 @@ open class DefaultMetrics(
         return motherboardMetrics
     }
 
-    override fun systemMetrics(): SystemMetrics {
-        return systemMetrics
+    override fun systemLoad(sort: ProcessSort, limit: Int): SystemLoad {
+        return SystemLoad(
+            cpuMetrics.uptime(),
+            cpuMetrics.cpuLoad().systemLoadAverage,
+            cpuMetrics.cpuLoad(),
+            networkMetrics.networkInterfaceLoads(),
+            networkMetrics.connectivity(),
+            diskMetrics.diskLoads(),
+            fileSystemMetrics.fileSystemLoads(),
+            memoryMetrics.memoryLoad(),
+            processesMetrics.processesInfo(sort, limit).processes,
+            gpuMetrics.gpuLoads(),
+            motherboardMetrics.motherboardHealth()
+        )
+    }
+
+    override fun systemInfo(): SystemInfo {
+        return SystemInfo(
+            hostName = EnvironmentUtils.hostName,
+            operatingSystem = operatingSystem,
+            platform = platform,
+            cpuInfo = cpuMetrics.cpuInfo(),
+            motherboard = motherboardMetrics.motherboard(),
+            memory = memoryMetrics.memoryInfo(),
+            fileSystems = fileSystemMetrics.fileSystems(),
+            networkInterfaces = networkMetrics.networkInterfaces(),
+            gpus = gpuMetrics.gpus()
+        )
     }
 }
