+++ src/main/kotlin/com/krillsson/sysapi/core/metrics/windows/WindowsMetrics.kt
@@ -1,5 +1,7 @@
 package com.krillsson.sysapi.core.metrics.windows
 
+import com.krillsson.sysapi.core.domain.system.OperatingSystem
+import com.krillsson.sysapi.core.domain.system.Platform
 import com.krillsson.sysapi.core.metrics.MemoryMetrics
 import com.krillsson.sysapi.core.metrics.SystemMetrics
 import com.krillsson.sysapi.core.metrics.defaultimpl.*
@@ -18,7 +20,8 @@ class WindowsMetrics(
     processesMetrics: DefaultProcessesMetrics,
     motherboardMetrics: WindowsMotherboardMetrics,
     memoryMetrics: MemoryMetrics,
-    systemMetrics: SystemMetrics,
+    operatingSystem: OperatingSystem,
+    platform: Platform,
     private val ohmManagerFactory: OHMManager
 ) : DefaultMetrics(
     cpuMetrics,
@@ -29,7 +32,8 @@ class WindowsMetrics(
     processesMetrics,
     motherboardMetrics,
     memoryMetrics,
-    systemMetrics
+    operatingSystem,
+    platform
 ) {
     override fun initialize() {
         super.initialize()
